@@ -1,63 +1,62 @@
 {
   "name": "@oclif/plugin-plugins",
   "description": "plugins plugin for oclif",
-  "version": "3.9.4",
+  "version": "3.9.4-qa.5",
   "author": "Salesforce",
   "bugs": "https://github.com/oclif/plugin-plugins/issues",
   "dependencies": {
-    "@oclif/core": "^2.15.0",
-    "chalk": "^4.1.2",
+    "@oclif/core": "^3.7.1",
+    "chalk": "^5.3.0",
     "debug": "^4.3.4",
-    "http-call": "^5.2.2",
-    "load-json-file": "^5.3.0",
     "npm": "9.8.1",
     "npm-run-path": "^4.0.1",
     "semver": "^7.5.4",
     "shelljs": "^0.8.5",
-    "tslib": "^2.6.2",
     "validate-npm-package-name": "^5.0.0",
     "yarn": "^1.22.18"
   },
   "devDependencies": {
-    "@commitlint/config-conventional": "^12.1.4",
-    "@oclif/plugin-help": "^5.2.20",
-    "@oclif/test": "^2.5.6",
-    "@types/chai": "^4.3.9",
-    "@types/glob": "^8.1.0",
-    "@types/mocha": "^8.2.3",
-    "@types/node": "^14.18.63",
+    "@commitlint/config-conventional": "^17.8.0",
+    "@oclif/plugin-help": "^6",
+    "@oclif/prettier-config": "^0.2.1",
+    "@types/chai": "^4.3.6",
+    "@types/debug": "^4.1.9",
+    "@types/mocha": "^10.0.2",
+    "@types/node": "^18",
     "@types/semver": "^7.5.3",
-    "@types/shelljs": "^0.8.14",
-    "@types/supports-color": "^7.2.0",
-    "@types/validate-npm-package-name": "^4.0.1",
+    "@types/shelljs": "^0.8.13",
+    "@types/sinon": "^10.0.19",
+    "@types/validate-npm-package-name": "^4.0.0",
     "chai": "^4.3.10",
-    "commitlint": "^12.1.4",
-    "eslint": "^7.31.0",
-    "eslint-config-oclif": "^4",
-    "eslint-config-oclif-typescript": "^1.0.3",
-    "fancy-test": "^2.0.35",
-    "globby": "^11.1.0",
-    "husky": "6",
-    "mocha": "^9.2.2",
-    "nock": "^13.3.6",
-    "oclif": "^3.9.1",
+    "commitlint": "^17.8.0",
+    "eslint": "^8.51.0",
+    "eslint-config-oclif": "^5.0.0",
+    "eslint-config-oclif-typescript": "^3.0.6",
+    "eslint-config-prettier": "^9.0.0",
+    "husky": "^8.0.3",
+    "lint-staged": "^15.0.1",
+    "mocha": "^10.2.0",
+    "oclif": "^4.0.2",
+    "prettier": "^3.0.3",
     "shx": "^0.3.4",
-    "ts-node": "^10.7.0",
-    "typescript": "4.9.5"
+    "sinon": "^16.1.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
   },
   "engines": {
-    "node": ">=16"
+    "node": ">=18.0.0"
   },
+  "exports": "./lib/index.js",
   "files": [
     "oclif.manifest.json",
-    "/lib"
+    "/lib",
+    "/oclif.lock"
   ],
   "homepage": "https://github.com/oclif/plugin-plugins",
   "keywords": [
     "oclif-plugin"
   ],
   "license": "MIT",
-  "main": "lib/index.js",
   "oclif": {
     "commands": "./lib/commands",
     "hooks": {
