@@ -1,12 +1,9 @@
-import {Command, Flags} from '@oclif/core'
+import {Command, Flags, ux} from '@oclif/core'
 
-import Plugins from '../../plugins'
+import Plugins from '../../plugins.js'
+import {YarnMessagesCache} from '../../util.js'
 
 export default class PluginsUpdate extends Command {
-  static topic = 'plugins'
-
-  static command = 'update'
-
   static description = 'Update installed plugins.'
 
   static flags = {
@@ -19,6 +16,12 @@ export default class PluginsUpdate extends Command {
   async run(): Promise<void> {
     const {flags} = await this.parse(PluginsUpdate)
     this.plugins.verbose = flags.verbose
+    ux.action.start(`${this.config.name}: Updating plugins`)
+
     await this.plugins.update()
+
+    ux.action.stop()
+
+    YarnMessagesCache.getInstance().flush()
   }
 }
