@@ -1,12 +1,14 @@
 import {Hook} from '@oclif/core'
 
-import Plugins from '../plugins'
+import Plugins from '../plugins.js'
 
 export const update: Hook<'update'> = async function () {
   const plugins = new Plugins(this.config)
   try {
     await plugins.update()
-  } catch (error: any) {
-    this.error(error.message)
+  } catch (error: unknown) {
+    if (error instanceof Error) {
+      this.error(error.message)
+    }
   }
 }
